fix: address review — split reward event types into AdRewardEvents.swift
Moves AdRewardFailureReason and the three AdReward* event structs out of
AdEvent.swift into a new file, and registers it in the Xcode project. The
file_length disable on AdEvent.swift stays since the remainder is still
above the 400-line limit.
